Minimal changes to get it compiling (and not crashing when ran under valgrind. Changes don't affect 32 bit mode). Changed aplWeaponClip and friends to int* ,but intptr_t* works as well for me. Final type will determined once the con code has been checked.
git-svn-id: https://svn.eduke32.com/eduke32@709 1a8010ca-5511-0410-912e-c29ae57300e0

+++ polymer/eduke32/source/actors.c
@@ -34,7 +34,7 @@ inline void updateinterpolations(void)  //Stick at beginning of domovethings
     for (;i>=0;i--) oldipos[i] = *curipos[i];
 }
 
-void setinterpolation(intptr_t *posptr)
+void setinterpolation(int *posptr)
 {
     int i=numinterpolations-1;
 
@@ -46,7 +46,7 @@ void setinterpolation(intptr_t *posptr)
     numinterpolations++;
 }
 
-void stopinterpolation(intptr_t *posptr)
+void stopinterpolation(int *posptr)
 {
     int i=numinterpolations-1;
